% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbmapdata-package.R
\docType{package}
\name{wbmapdata-package}
\alias{wbmapdata}
\alias{wbmapdata-package}
\title{wbmapdata: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Pete Dodd \email{p.j.dodd@sheffield.ac.uk}

}
\keyword{internal}
